Update dividing AuthenticatorAlgorithm class
